requirements.txt
psycopg2-binary==2.9.7
gitpython==3.1.40
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.21
google-generativeai==0.3.2
